hereyouare/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
orjson
httpx
Pillow
langfuse
redis>=5.0
xxhash>=3.4
pytest>=7.0
pytest-asyncio>=0.21